chore: reorganize the files
