+++ e-token/src/processor/initialize_shuttle_ephemeral_ata.rs
@@ -1,16 +1,7 @@
-use crate::processor::initialize_ephemeral_ata::initialize_ephemeral_ata_with_sponsor;
 use core::marker::PhantomData;
-use ephemeral_spl_api::state::{load_initialized, load_mut, RawType};
-use pinocchio::sysvars::rent::Rent;
-use pinocchio::sysvars::Sysvar;
-use pinocchio::{address::address_eq, cpi::Signer};
-use pinocchio_system::instructions::{CreateAccount, Transfer};
-use {
-    ephemeral_spl_api::state::shuttle_ephemeral_ata::ShuttleMetadata,
-    pinocchio::{error::ProgramError, AccountView, ProgramResult},
-};
+use pinocchio::{error::ProgramError, AccountView, ProgramResult};
 
-const SHUTTLE_METADATA_V0_LEN: usize = 68;
+use crate::processor::internal::ephemeral_ata::initialize_shuttle_ephemeral_ata_with_sponsor;
 
 #[inline(always)]
 pub fn process_initialize_shuttle_ephemeral_ata(
@@ -52,125 +43,6 @@ pub fn process_initialize_shuttle_ephemeral_ata(
         system_program_info,
         args.shuttle_id(),
     )?;
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-#[inline(never)]
-pub(crate) fn initialize_shuttle_ephemeral_ata_with_sponsor(
-    sponsor_info: &AccountView,
-    sponsor_signer: Option<Signer<'_, '_>>,
-    shuttle_info: &AccountView,
-    shuttle_eata_info: &AccountView,
-    shuttle_wallet_ata_info: &AccountView,
-    refund_recipient_info: &AccountView,
-    owner_info: &AccountView,
-    mint_info: &AccountView,
-    token_program_info: &AccountView,
-    system_program_info: &AccountView,
-    shuttle_id: u32,
-) -> ProgramResult {
-    let shuttle_id_seed = shuttle_id.to_le_bytes();
-    let (derived_shuttle_pda, shuttle_bump) =
-        ShuttleMetadata::find_pda(owner_info.address(), mint_info.address(), shuttle_id);
-    if !address_eq(&derived_shuttle_pda, shuttle_info.address()) {
-        return Err(ProgramError::InvalidSeeds);
-    }
-
-    let shuttle_is_owned_by_program = unsafe { shuttle_info.owner().eq(&crate::ID) };
-
-    if !shuttle_is_owned_by_program {
-        let bump = [shuttle_bump];
-        let shuttle_seed = ShuttleMetadata::signer_seeds(
-            owner_info.address(),
-            mint_info.address(),
-            &shuttle_id_seed,
-            &bump,
-        );
-        let shuttle_signer = Signer::from(&shuttle_seed);
-
-        let create_shuttle = CreateAccount {
-            from: sponsor_info,
-            to: shuttle_info,
-            space: ShuttleMetadata::LEN as u64,
-            lamports: Rent::get()?.try_minimum_balance(ShuttleMetadata::LEN)?,
-            owner: &crate::ID,
-        };
-        if let Some(sponsor_signer) = sponsor_signer.as_ref() {
-            let signers = [sponsor_signer.clone(), shuttle_signer];
-            create_shuttle.invoke_signed(&signers)?;
-        } else {
-            let signers = [shuttle_signer];
-            create_shuttle.invoke_signed(&signers)?;
-        }
-
-        let shuttle = unsafe { load_mut::<ShuttleMetadata>(shuttle_info.borrow_unchecked_mut())? };
-
-        shuttle.owner = *owner_info.address();
-        shuttle.payer = *refund_recipient_info.address();
-        shuttle.id = shuttle_id;
-        shuttle.bump = shuttle_bump;
-    } else {
-        // Migrate legacy shuttle metadata
-        if shuttle_info.data_len() == SHUTTLE_METADATA_V0_LEN {
-            let current_lamports = shuttle_info.lamports();
-            if current_lamports < Rent::get()?.try_minimum_balance(ShuttleMetadata::LEN)? {
-                if let Some(sponsor_signer) = sponsor_signer.clone() {
-                    Transfer {
-                        from: sponsor_info,
-                        to: shuttle_info,
-                        lamports: Rent::get()?.try_minimum_balance(ShuttleMetadata::LEN)?
-                            - current_lamports,
-                    }
-                    .invoke_signed(&[sponsor_signer])?;
-                } else {
-                    Transfer {
-                        from: sponsor_info,
-                        to: shuttle_info,
-                        lamports: Rent::get()?.try_minimum_balance(ShuttleMetadata::LEN)?
-                            - current_lamports,
-                    }
-                    .invoke()?;
-                }
-            }
-            shuttle_info.resize(ShuttleMetadata::LEN)?;
-            let shuttle =
-                load_mut::<ShuttleMetadata>(unsafe { shuttle_info.borrow_unchecked_mut() })?;
-            shuttle.bump = shuttle_bump;
-        }
-
-        let shuttle =
-            load_initialized::<ShuttleMetadata>(unsafe { shuttle_info.borrow_unchecked() })?;
-        if shuttle.id != shuttle_id
-            || !address_eq(&shuttle.owner, owner_info.address())
-            || !address_eq(&shuttle.payer, refund_recipient_info.address())
-        {
-            return Err(ProgramError::InvalidAccountData);
-        }
-    }
-
-    initialize_ephemeral_ata_with_sponsor(
-        shuttle_eata_info,
-        sponsor_info,
-        sponsor_signer.clone(),
-        shuttle_info,
-        mint_info,
-    )?;
-
-    let ata_ix = pinocchio_associated_token_account::instructions::CreateIdempotent {
-        funding_account: sponsor_info,
-        account: shuttle_wallet_ata_info,
-        wallet: shuttle_info,
-        mint: mint_info,
-        system_program: system_program_info,
-        token_program: token_program_info,
-    };
-    if let Some(sponsor_signer) = sponsor_signer {
-        ata_ix.invoke_signed(&[sponsor_signer])?;
-    } else {
-        ata_ix.invoke()?;
-    }
 
     Ok(())
 }
